Move map components into root components directory
