+++ components/molecules/marker-list/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import PopupMarker from '_components/atoms/PopupMarker'
+import PopupMarker from '_components/atoms/popup-marker'
 
 const MarkerList = ({ markers }) => {
   const items = markers.map(({ key, ...props }) => <PopupMarker key={key} {...props} />)
